Applied another patch by Bart which fixes a compile warning on 64Bit architectures and added the format flag %X (for the album URL) to meta().
git-svn-id: svn://nex.scrapping.cc/shell-fm/trunk@59 7df44517-d413-0410-91cf-82ca28b36b55

+++ http.c
@@ -64,7 +64,7 @@ char ** fetch(char * const url, FILE ** pHandle, const char * data) {
 	fprintf(fd, headFormat, data ? "POST" : "GET", file ? file : "", host);
 
 	if(data)
-		fprintf(fd, "Content-Length: %d\r\n\r\n%s\r\n", strlen(data), data);
+		fprintf(fd, "Content-Length: %ld\r\n\r\n%s\r\n", (long) strlen(data), data);
 
 	fputs("\r\n", fd);
 	fflush(fd);
